requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-haystack==3.3.0
elasticsearch==8.15.1
djangorestframework-simplejwt==5.3.1